bootcamp-AndreSioneck/bootcamp-data-platform/requirements.txt
aws-cdk-lib>=2.240.0,<3.0.0
constructs>=10.5.0,<11.0.0
pre-commit>=3.0.0,<4.0.0
fake-web-events
boto3
psycopg2-binary